CB-11565 Extract inner Action classes
Housekeeping
