+++ environment/src/main/java/com/sequenceiq/environment/environment/flow/creation/AbstractEnvironmentCreationAction.java
@@ -1,0 +1,46 @@
+package com.sequenceiq.environment.environment.flow.creation;
+
+import java.util.Map;
+import java.util.Optional;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.statemachine.StateContext;
+
+import com.sequenceiq.cloudbreak.common.event.ResourceCrnPayload;
+import com.sequenceiq.cloudbreak.logger.MdcContext;
+import com.sequenceiq.environment.environment.flow.creation.event.EnvCreationStateSelectors;
+import com.sequenceiq.flow.core.AbstractAction;
+import com.sequenceiq.flow.core.CommonContext;
+import com.sequenceiq.flow.core.FlowParameters;
+
+abstract class AbstractEnvironmentCreationAction<P extends ResourceCrnPayload>
+        extends AbstractAction<EnvCreationState, EnvCreationStateSelectors, CommonContext, P> {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(AbstractEnvironmentCreationAction.class);
+
+    protected AbstractEnvironmentCreationAction(Class<P> payloadClass) {
+        super(payloadClass);
+    }
+
+    @Override
+    protected CommonContext createFlowContext(FlowParameters flowParameters, StateContext<EnvCreationState, EnvCreationStateSelectors> stateContext,
+            P payload) {
+        return new CommonContext(flowParameters);
+    }
+
+    @Override
+    protected Object getFailurePayload(P payload, Optional<CommonContext> flowContext, Exception ex) {
+        LOGGER.debug("Received failure payload: {}", payload);
+        return payload;
+    }
+
+    @Override
+    protected void prepareExecution(P payload, Map<Object, Object> variables) {
+        if (payload != null) {
+            MdcContext.builder().resourceCrn(payload.getResourceCrn()).buildMdc();
+        } else {
+            LOGGER.warn("Payload was null in prepareExecution so resourceCrn cannot be added to the MdcContext!");
+        }
+    }
+}
