+++ environment/src/main/java/com/sequenceiq/environment/environment/flow/deletion/AbstractEnvDeleteAction.java
@@ -1,0 +1,53 @@
+package com.sequenceiq.environment.environment.flow.deletion;
+
+import java.util.Map;
+import java.util.Optional;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.statemachine.StateContext;
+
+import com.sequenceiq.cloudbreak.common.event.ResourceCrnPayload;
+import com.sequenceiq.cloudbreak.logger.MdcContext;
+import com.sequenceiq.cloudbreak.util.NullUtil;
+import com.sequenceiq.environment.environment.flow.deletion.event.EnvDeleteFailedEvent;
+import com.sequenceiq.environment.environment.flow.deletion.event.EnvDeleteStateSelectors;
+import com.sequenceiq.flow.core.AbstractAction;
+import com.sequenceiq.flow.core.CommonContext;
+import com.sequenceiq.flow.core.FlowParameters;
+
+abstract class AbstractEnvDeleteAction<P extends ResourceCrnPayload>
+        extends AbstractAction<EnvDeleteState, EnvDeleteStateSelectors, CommonContext, P> {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(AbstractEnvDeleteAction.class);
+
+    protected AbstractEnvDeleteAction(Class<P> payloadClass) {
+        super(payloadClass);
+    }
+
+    @Override
+    protected CommonContext createFlowContext(FlowParameters flowParameters, StateContext<EnvDeleteState, EnvDeleteStateSelectors> stateContext,
+            P payload) {
+        return new CommonContext(flowParameters);
+    }
+
+    @Override
+    protected Object getFailurePayload(P payload, Optional<CommonContext> flowContext, Exception ex) {
+        LOGGER.debug("Received failure payload: {}", payload);
+        return EnvDeleteFailedEvent.builder()
+                .withException(ex)
+                .withEnvironmentId(NullUtil.getIfNotNullOtherwise(payload, ResourceCrnPayload::getResourceId, -1L))
+                .withResourceCrn(NullUtil.getIfNotNullOtherwise(payload, ResourceCrnPayload::getResourceCrn, "null CRN"))
+                .build();
+    }
+
+    @Override
+    protected void prepareExecution(P payload, Map<Object, Object> variables) {
+        if (payload != null) {
+            MdcContext.builder().resourceCrn(payload.getResourceCrn()).buildMdc();
+        } else {
+            LOGGER.warn("Payload was null in prepareExecution so resourceCrn cannot be added to the MdcContext!");
+        }
+    }
+
+}
