+++ environment/src/main/java/com/sequenceiq/environment/environment/flow/deletion/EnvClustersDeleteActions.java
@@ -6,7 +6,6 @@ import static com.sequenceiq.environment.environment.flow.deletion.event.EnvDele
 import static com.sequenceiq.environment.environment.flow.deletion.event.EnvDeleteHandlerSelectors.DELETE_EXPERIENCE_EVENT;
 
 import java.util.Map;
-import java.util.Optional;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -15,9 +14,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.statemachine.StateContext;
 import org.springframework.statemachine.action.Action;
 
-import com.sequenceiq.cloudbreak.common.event.ResourceCrnPayload;
 import com.sequenceiq.cloudbreak.event.ResourceEvent;
-import com.sequenceiq.cloudbreak.logger.MdcContext;
 import com.sequenceiq.environment.environment.EnvironmentStatus;
 import com.sequenceiq.environment.environment.domain.Environment;
 import com.sequenceiq.environment.environment.dto.EnvironmentDeletionDto;
@@ -30,7 +27,6 @@ import com.sequenceiq.environment.environment.service.EnvironmentStatusUpdateSer
 import com.sequenceiq.environment.events.EventSenderService;
 import com.sequenceiq.environment.metrics.EnvironmentMetricService;
 import com.sequenceiq.environment.metrics.MetricType;
-import com.sequenceiq.flow.core.AbstractAction;
 import com.sequenceiq.flow.core.CommonContext;
 import com.sequenceiq.flow.core.Flow;
 import com.sequenceiq.flow.core.FlowParameters;
@@ -169,35 +165,6 @@ public class EnvClustersDeleteActions {
                 .build();
 
         return environmentDeletionDto;
-    }
-
-    private abstract static class AbstractEnvClustersDeleteAction<P extends ResourceCrnPayload>
-            extends AbstractAction<EnvClustersDeleteState, EnvClustersDeleteStateSelectors, CommonContext, P> {
-
-        protected AbstractEnvClustersDeleteAction(Class<P> payloadClass) {
-            super(payloadClass);
-        }
-
-        @Override
-        protected CommonContext createFlowContext(FlowParameters flowParameters,
-                StateContext<EnvClustersDeleteState, EnvClustersDeleteStateSelectors> stateContext,
-                P payload) {
-            return new CommonContext(flowParameters);
-        }
-
-        @Override
-        protected Object getFailurePayload(P payload, Optional<CommonContext> flowContext, Exception ex) {
-            return payload;
-        }
-
-        @Override
-        protected void prepareExecution(P payload, Map<Object, Object> variables) {
-            if (payload != null) {
-                MdcContext.builder().resourceCrn(payload.getResourceCrn()).buildMdc();
-            } else {
-                LOGGER.warn("Payload was null in prepareExecution so resourceCrn cannot be added to the MdcContext!");
-            }
-        }
     }
 
 }
